Addressed comments from David and Almann
