+++ lang/test/org/partiql/lang/eval/builtins/MakeDateEvaluationTest.kt
@@ -1,16 +1,20 @@
 package org.partiql.lang.eval.builtins
 
 import com.amazon.ion.IonStruct
+import junitparams.Parameters
+import org.assertj.core.api.Assertions
+import org.junit.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ArgumentsSource
 import org.partiql.lang.errors.ErrorCode
-import org.partiql.lang.eval.EvaluationException
-import org.partiql.lang.eval.EvaluatorTestBase
-import org.partiql.lang.eval.ExprValueType
+import org.partiql.lang.eval.*
 import org.partiql.lang.util.ArgumentsProviderBase
 import org.partiql.lang.util.get
+import java.time.LocalDate
 
 class MakeDateEvaluationTest : EvaluatorTestBase() {
+    private val env = Environment.standard()
+    private val subject = MakeDateExprFunction(valueFactory)
 
     data class MakeDateTestCase(val query: String, val expected: String?, val errorCode: ErrorCode?)
 
@@ -68,5 +72,69 @@ class MakeDateEvaluationTest : EvaluatorTestBase() {
             case("MAKE_DATE(2021,02,'twenty-seven')", ErrorCode.EVALUATOR_INCORRECT_TYPE_OF_ARGUMENTS_TO_FUNC_CALL),
             case("MAKE_DATE('2021',02,27)", ErrorCode.EVALUATOR_INCORRECT_TYPE_OF_ARGUMENTS_TO_FUNC_CALL)
         )
+    }
+
+    private fun callMakeDate(vararg args: Any): LocalDate? {
+        val value = subject.call(env, args.map { anyToExprValue(it) }.toList())
+        return when(value.type) {
+            ExprValueType.NULL -> null
+            else -> value.dateValue()
+        }
+    }
+
+    @Test
+    fun lessArguments() {
+        Assertions.assertThatThrownBy { callMakeDate(2021) }
+            .hasMessage("make_date takes exactly 3 arguments, received: 1")
+            .isExactlyInstanceOf(EvaluationException::class.java)
+    }
+
+    @Test
+    fun lessArguments2() {
+        Assertions.assertThatThrownBy { callMakeDate(2021, 2) }
+            .hasMessage("make_date takes exactly 3 arguments, received: 2")
+            .isExactlyInstanceOf(EvaluationException::class.java)
+    }
+
+    @Test
+    fun moreArguments() {
+        Assertions.assertThatThrownBy { callMakeDate(2021, 2, 29, 1.123) }
+            .hasMessage("make_date takes exactly 3 arguments, received: 4")
+            .isExactlyInstanceOf(EvaluationException::class.java)
+    }
+
+    @Test
+    fun wrongTypeOfArgumentForYear() {
+        Assertions.assertThatThrownBy { callMakeDate("2021", 2, 28) }
+            .hasMessage("Invalid argument type for make_date")
+            .isExactlyInstanceOf(EvaluationException::class.java)
+    }
+
+    @Test
+    fun wrongTypeOfArgumentForMonth() {
+        Assertions.assertThatThrownBy { callMakeDate(2021, 2.0, 28) }
+            .hasMessage("Invalid argument type for make_date")
+            .isExactlyInstanceOf(EvaluationException::class.java)
+    }
+
+    @Test
+    fun wrongTypeOfArgumentForDay() {
+        Assertions.assertThatThrownBy { callMakeDate(2021, 2, "twenty-eight") }
+            .hasMessage("Invalid argument type for make_date")
+            .isExactlyInstanceOf(EvaluationException::class.java)
+    }
+
+    fun parametersForMakeDate(): List<Pair<LocalDate?, () -> LocalDate?>> = listOf(
+        LocalDate.of(2021, 3,26) to {callMakeDate(2021, 3, 26)},
+        LocalDate.of(2021, 2,28) to {callMakeDate(2021, 2, 28)},
+        LocalDate.of(2020, 2,29) to {callMakeDate(2020, 2, 29)}
+    )
+
+    @Test
+    @Parameters
+    fun makeDate(params: Pair<LocalDate?, () -> LocalDate?>) {
+        val (expected, call) = params
+
+        assertEquals(expected, call.invoke())
     }
 }
